Move all authentication related code from the Main activity into the Base activity so it can be used throughout the app
